requirements.txt
chardet==5.2.0
cloudpickle==3.1.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.45.1
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpld3==0.5.9
natsort==8.4.0
networkx==3.2.1
numpy==1.24.4
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
psutil==5.9.6
pycparser==2.22
pyelftools==0.30
pygad==3.3.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
sark==8.2.0
scikit-learn==1.0.2
scipy==1.9.1
seaborn==0.13.0
setproctitle==1.3.3
six==1.16.0
threadpoolctl==3.2.0
tqdm==4.66.1
tree-sitter==0.22.3
tree-sitter-c==0.21.3
tzdata==2023.3
wrapt==1.16.0
xgboost==2.0.2
yara-python==4.4.0
zipp==3.17.0
